chore: add car search engine project
This project simulates a website where the user can filter by different
fields such as:

Brand, price, color, price, date...

The query result is updated as soon as the user filters by any field.
If the result is empty, a message is shown to notify the user.

The data related to the cars is stored as a json object in db.js.

Missing README.md file, will be added in the future.